refactor(metadata,swagger): finalize hexagonal layering and align docs
