+++ packages/metadata/src/core/utils/object.ts
