+++ packages/swagger/package.json
@@ -28,10 +28,6 @@
         ".": {
             "types": "./dist/index.d.mts",
             "import": "./dist/index.mjs"
-        },
-        "./metadata": {
-            "types": "./dist/metadata.d.mts",
-            "import": "./dist/metadata.mjs"
         }
     },
     "files": [
