--- packages/swagger/src/metadata.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2021.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-export * from '@trapi/metadata';
