+++ packages/swagger/tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsdown';
 
 export default defineConfig({
-    entry: ['src/index.ts', 'src/metadata.ts'],
+    entry: ['src/index.ts'],
     format: 'esm',
     dts: true,
     sourcemap: true,
